fix: extract calc modal to own comp, impl data flow for calc
