+++ src/__demo__/CalculationModal.stories.js
@@ -1,0 +1,7 @@
+import { storiesOf } from '@storybook/react'
+import React from 'react'
+import CalculationModal from '../components/DataDimension/CalculationModal.js'
+
+storiesOf('CalculationModal', module).add('Default', () => {
+    return <CalculationModal />
+})
