--- src/__demo__/EditCalculation.stories.js
@@ -1,7 +1,0 @@
-import { storiesOf } from '@storybook/react'
-import React from 'react'
-import EditCalculation from '../components/DataDimension/EditCalculation.js'
-
-storiesOf('EditCalculation', module).add('Default', () => {
-    return <EditCalculation />
-})
